services/yolo/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
ultralytics==8.3.43
pillow==10.4.0
requests==2.32.3
python-multipart==0.0.9
